+++ src/episodeChecker.ts
@@ -1,5 +1,5 @@
 import { type EpisodeData } from './types.js';
-import { resolvePlaylistId, isPrivateVideo } from './postUtils.js';
+import { resolvePlaylistId } from './postUtils.js';
 
 const YT_API_BASE = 'https://youtube.googleapis.com/youtube/v3';
 
@@ -62,13 +62,6 @@ export async function fetchPlaylistVideos(
         link: `https://www.youtube.com/watch?v=${videoId}`,
         episodeNumber: undefined,
       };
-    })
-    .filter((video) => {
-      if (isPrivateVideo(video)) {
-        console.log(`[episodeChecker] Ignoring private/deleted video (${video.guid})`);
-        return false;
-      }
-      return true;
     });
 }